move to use Cobra
