+++ .gitignore
@@ -20,3 +20,5 @@
 # Go workspace file
 go.work
 go.work.sum
+
+gh-commit-remap
