+++ go.mod
@@ -1,3 +1,10 @@
 module github.com/kuhlman-labs/gh-commit-remap
 
 go 1.22.5
+
+require github.com/spf13/cobra v1.8.1
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
